Add an unsubscribe option
It's helpful to me to have an un-subscribe option since I am using promises to catch returns (think ping/pong) based on a user-action.  The user clicks a button and I use mqtt to find out if the service is awake before I send the request -- otherwise I handle the click differently.  I want to unsubscribe if the service is there or not there so that the next time around I can do it again.  The mqtt topic is different and auto-generated so I don't want to leave subscriptions hanging around.

and it would be a bother to have to